+++ src/components/post/PostForms.tsx
@@ -5,8 +5,8 @@ import HashtagInput from '../ui/HashtagInput';
 import TimeInputWithIcon from '../ui/TimeInputWithIcon';
 import PostAddress from './PostAddress';
 import useMediaQuery from '@/hooks/byUse/useMediaQuery';
-import { usePostForm } from '@/hooks/byUse/usePostForm';
 import { useSubmitForm } from '@/hooks/queries/post/useFormMutations';
+import { usePostForm } from '@/hooks/useCustomForm/usePostForm';
 import { IconPluslg } from '@/lib/icon/Icon_Plus_lg';
 import { formSchema } from '@/schemas/formSchemas';
 import { saveTags, updateImagePostId } from '@/services/server-action/formActions';
